--- a/jwe/internal/cmd/genheader/main.go
+++ b/jwe/internal/cmd/genheader/main.go
@@ -101,7 +101,7 @@ func generateHeaders() error {
 			typ:    `jwa.KeyEncryptionAlgorithm`,
 			key:    `alg`,
 			//			comment:   `https://tools.ietf.org/html/rfc7515#section-4.1.1`,
-			jsonTag: "`" + `json:"alg,omitempty"` + "`",
+			jsonTag:  "`" + `json:"alg,omitempty"` + "`",
 		},
 		{
 			name:   `compression`,
@@ -434,64 +434,86 @@ func generateHeaders() error {
 	fmt.Fprintf(&buf, "\n}")
 
 	fmt.Fprintf(&buf, "\n\nfunc (h *stdHeaders) UnmarshalJSON(buf []byte) error {")
-	fmt.Fprintf(&buf, "\nh.mu.Lock()")
-	fmt.Fprintf(&buf, "\ndefer h.mu.Unlock()")
-	fmt.Fprintf(&buf, "\nvar proxy standardHeadersMarshalProxy")
-	fmt.Fprintf(&buf, "\nif err := json.Unmarshal(buf, &proxy); err != nil {")
-	fmt.Fprintf(&buf, "\nreturn errors.Wrap(err, `failed to unmarshal headers`)")
-	fmt.Fprintf(&buf, "\n}")
-
-	// Copy every field except for jwk and ephemeralPublicKey, whose type needs to be guessed
-	fmt.Fprintf(&buf, "\n\nh.jwk = nil")
-	fmt.Fprintf(&buf, "\nif jwkField := proxy.Xjwk; len(jwkField) > 0 {")
-	fmt.Fprintf(&buf, "\njwkKey, err := jwk.ParseKey([]byte(proxy.Xjwk))")
-	fmt.Fprintf(&buf, "\n if err != nil {")
-	fmt.Fprintf(&buf, "\nreturn errors.Wrap(err, `failed to parse jwk field`)")
-	fmt.Fprintf(&buf, "\n}")
-	fmt.Fprintf(&buf, "\nh.jwk = jwkKey")
-	fmt.Fprintf(&buf, "\n}")
-
-	fmt.Fprintf(&buf, "\n\nh.ephemeralPublicKey = nil")
-	fmt.Fprintf(&buf, "\nif epkField := proxy.XephemeralPublicKey; len(epkField) > 0 {")
-	fmt.Fprintf(&buf, "\nepk, err := jwk.ParseKey([]byte(proxy.XephemeralPublicKey))")
-	fmt.Fprintf(&buf, "\nif err != nil {")
-	fmt.Fprintf(&buf, "\nreturn errors.Wrap(err, `failed to parse epk field`)")
-	fmt.Fprintf(&buf, "\n}")
-	fmt.Fprintf(&buf, "\nh.ephemeralPublicKey = epk")
-	fmt.Fprintf(&buf, "\n}")
-
-	for _, name := range []string{agreementPartyUInfo, agreementPartyVInfo} {
-		fmt.Fprintf(&buf, "\n\nh.%s = nil", name)
-		fmt.Fprintf(&buf, "\nif v := proxy.X%s; len(v) > 0 {", name)
-		fmt.Fprintf(&buf, "\ndecoded, err := base64.Decode(v)")
-		fmt.Fprintf(&buf, "\nif err != nil {")
-		fmt.Fprintf(&buf, "\nreturn errors.Wrap(err, `failed to decode base64`)")
-		fmt.Fprintf(&buf, "\n}")
-		fmt.Fprintf(&buf, "\nh.%s = decoded", name)
-		fmt.Fprintf(&buf, "\n}")
+	for _, f := range fields {
+		fmt.Fprintf(&buf, "\nh.%s = nil", f.name)
 	}
 
+	fmt.Fprintf(&buf, "\ndec := json.NewDecoder(bytes.NewReader(buf))")
+	fmt.Fprintf(&buf, "\nLOOP:")
+	fmt.Fprintf(&buf, "\nfor {")
+	fmt.Fprintf(&buf, "\ntok, err := dec.Token()")
+	fmt.Fprintf(&buf, "\nif err != nil {")
+	fmt.Fprintf(&buf, "\nreturn errors.Wrap(err, `error reading token`)")
+	fmt.Fprintf(&buf, "\n}")
+	fmt.Fprintf(&buf, "\nswitch tok := tok.(type) {")
+	fmt.Fprintf(&buf, "\ncase json.Delim:")
+	fmt.Fprintf(&buf, "\n// Assuming we're doing everything correctly, we should ONLY")
+	fmt.Fprintf(&buf, "\n// get either '{' or '}' here.")
+	fmt.Fprintf(&buf, "\nif tok == '}' { // End of object")
+	fmt.Fprintf(&buf, "\nbreak LOOP")
+	fmt.Fprintf(&buf, "\n} else if tok != '{' {")
+	fmt.Fprintf(&buf, "\nreturn errors.Errorf(`expected '{', but got '%%c'`, tok)")
+	fmt.Fprintf(&buf, "\n}")
+	fmt.Fprintf(&buf, "\ncase string: // Objects can only have string keys")
+	fmt.Fprintf(&buf, "\nswitch tok {")
+
 	for _, f := range fields {
-		switch f.name {
-		case jwkKey, ephemeralPublicKey, agreementPartyUInfo, agreementPartyVInfo:
-			continue
-		default:
-			fmt.Fprintf(&buf, "\nh.%[1]s = proxy.X%[1]s", f.name)
+		if f.typ == "string" {
+			fmt.Fprintf(&buf, "\ncase %sKey:", f.method)
+			fmt.Fprintf(&buf, "\nif err := json.AssignNextStringToken(&h.%s, dec); err != nil {", f.name)
+			fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to decode value for key %%s`, %sKey)", f.method)
+			fmt.Fprintf(&buf, "\n}")
+		} else if f.typ == "[]byte" {
+			name := f.method
+			fmt.Fprintf(&buf, "\ncase %sKey:", name)
+			fmt.Fprintf(&buf, "\nif err := json.AssignNextBytesToken(&h.%s, dec); err != nil {", f.name)
+			fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to decode value for key %%s`, %sKey)", name)
+			fmt.Fprintf(&buf, "\n}")
+		} else if f.typ == "jwk.Key" {
+			name := f.method
+			fmt.Fprintf(&buf, "\ncase %sKey:", name)
+			fmt.Fprintf(&buf, "\nvar buf json.RawMessage")
+			fmt.Fprintf(&buf, "\nif err := dec.Decode(&buf); err != nil {")
+			fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to decode value for key %%s`, %sKey)", name)
+			fmt.Fprintf(&buf, "\n}")
+			fmt.Fprintf(&buf, "\nkey, err := jwk.ParseKey(buf)")
+			fmt.Fprintf(&buf, "\nif err != nil {")
+			fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to parse JWK for key %%s`, %sKey)", name)
+			fmt.Fprintf(&buf, "\n}")
+			fmt.Fprintf(&buf, "\nh.%s = key", f.name)
+		} else if strings.HasPrefix(f.typ, "[]") {
+			name := f.method
+			fmt.Fprintf(&buf, "\ncase %sKey:", name)
+			fmt.Fprintf(&buf, "\nvar decoded %s", f.typ)
+			fmt.Fprintf(&buf, "\nif err := dec.Decode(decoded); err != nil {")
+			fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to decode value for key %%s`, %sKey)", name)
+			fmt.Fprintf(&buf, "\n}")
+			fmt.Fprintf(&buf, "\nh.%s = decoded", f.name)
+		} else {
+			name := f.method
+			fmt.Fprintf(&buf, "\ncase %sKey:", name)
+			fmt.Fprintf(&buf, "\nvar decoded %s", f.typ)
+			fmt.Fprintf(&buf, "\nif err := dec.Decode(&decoded); err != nil {")
+			fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to decode value for key %%s`, %sKey)", name)
+			fmt.Fprintf(&buf, "\n}")
+			fmt.Fprintf(&buf, "\nh.%s = &decoded", f.name)
 		}
 	}
-
-	// Now for the fun part... It's quite silly, but we need to check if we
-	// have other parameters.
-	fmt.Fprintf(&buf, "\nvar m map[string]interface{}")
-	fmt.Fprintf(&buf, "\nif err := json.Unmarshal(buf, &m); err != nil {")
-	fmt.Fprintf(&buf, "\nreturn errors.Wrap(err, `failed to parse privsate parameters`)")
+	fmt.Fprintf(&buf, "\ndefault:")
+	fmt.Fprintf(&buf, "\nvar decoded interface{}")
+	fmt.Fprintf(&buf, "\nif err := dec.Decode(&decoded); err != nil {")
+	fmt.Fprintf(&buf, "\nreturn errors.Wrapf(err, `failed to decode field %%s`, tok)")
 	fmt.Fprintf(&buf, "\n}")
-	// Delete all known keys
-	for _, f := range fields {
-		fmt.Fprintf(&buf, "\ndelete(m, %sKey)", f.method)
-	}
+	fmt.Fprintf(&buf, "\nif h.privateParams == nil {")
+	fmt.Fprintf(&buf, "\nh.privateParams = make(map[string]interface{})")
+	fmt.Fprintf(&buf, "\n}")
+	fmt.Fprintf(&buf, "\nh.privateParams[tok] = decoded")
+	fmt.Fprintf(&buf, "\n}")
+	fmt.Fprintf(&buf, "\ndefault:")
+	fmt.Fprintf(&buf, "\nreturn errors.Errorf(`invalid token %%T`, tok)")
+	fmt.Fprintf(&buf, "\n}")
+	fmt.Fprintf(&buf, "\n}")
 
-	fmt.Fprintf(&buf, "\nh.privateParams = m")
 	fmt.Fprintf(&buf, "\nreturn nil")
 	fmt.Fprintf(&buf, "\n}")