As of b08eca9fcd240e5ad47d4c42bdf5624dc7d2605b, we are using the streaming interface for encoding/json, and abolished uses of proxy objects in some cases.

In general parsing/serializing is about 10%~20% faster, but uses more memory, due to the way 
 `(encoding/json).Decoder.Token()`  is implemented.

There is a plan to incorporate github.com/goccy/go-json for even more performance improvements

```
Benchmark comparison for:
  develop/v1.1.0
  topic/v1.1.0-marshal-optimize

name                                                      old time/op    new time/op    delta
JWE/Serialization/JSON/json.Marshal-8                       13.5µs ± 4%    13.0µs ± 1%   -3.82%  (p=0.032 n=5+5)
JWE/Serialization/JSON/json.Unmarshal-8                     3.41µs ± 2%    3.42µs ± 1%     ~     (p=0.690 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.Parse-8                 15.5µs ± 9%    12.7µs ± 2%  -17.84%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseString-8           14.7µs ± 1%    12.8µs ± 0%  -12.91%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseReader-8           17.3µs ± 7%    15.1µs ± 1%  -12.63%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/json.Marshal-8              6.08µs ± 2%    5.17µs ± 2%  -14.91%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.Parse-8                55.2µs ± 1%    47.9µs ± 5%  -13.22%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseString-8          55.1µs ± 1%    46.8µs ± 2%  -15.01%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseReader-8          58.9µs ± 2%    50.6µs ± 1%  -14.23%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/json.Marshal-8             18.9µs ± 1%    16.4µs ± 2%  -13.37%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.Parse-8                  11.4µs ± 1%    10.4µs ± 2%   -8.98%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseString-8            11.5µs ± 1%    10.5µs ± 1%   -8.76%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseReader-8            13.8µs ± 1%    12.7µs ± 3%   -8.03%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/json.Marshal-8               6.48µs ± 1%    5.92µs ± 5%   -8.64%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.Parse-8                 14.9µs ± 1%    13.3µs ± 1%  -11.01%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseString-8           15.2µs ± 5%    13.4µs ± 1%  -11.75%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseReader-8           17.4µs ± 1%    15.8µs ± 2%   -9.12%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/json.Marshal-8              7.96µs ± 1%    7.27µs ± 4%   -8.62%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.Parse-8           7.38µs ± 1%    6.62µs ± 1%  -10.37%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseString-8     7.43µs ± 1%    6.67µs ± 3%  -10.14%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseReader-8     9.63µs ± 3%    8.66µs ± 2%   -9.99%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/json.Marshal-8        3.57µs ± 1%    3.46µs ± 9%     ~     (p=0.690 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.Parse-8          7.41µs ± 1%    6.60µs ± 2%  -11.00%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseString-8    7.43µs ± 0%    6.68µs ± 2%  -10.09%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseReader-8    9.49µs ± 2%    8.69µs ± 4%   -8.44%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal-8       3.56µs ± 1%    3.24µs ± 1%   -8.99%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.Parse-8                       3.20µs ± 2%    2.48µs ± 1%  -22.37%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseString-8                 3.23µs ± 1%    2.54µs ± 1%  -21.35%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseReader-8                 5.28µs ± 4%    4.40µs ± 2%  -16.58%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.Parse-8                          21.8µs ± 7%    18.1µs ± 1%  -16.97%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseString-8                    21.3µs ± 1%    18.5µs ± 2%  -13.38%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseReader-8                    24.9µs ± 1%    22.0µs ± 2%  -11.65%  (p=0.008 n=5+5)
JWS/Serialization/JSON/json.Marshal-8                       12.9µs ± 1%     9.8µs ± 2%  -24.17%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseString-8                    7.50µs ± 1%    6.17µs ± 1%  -17.68%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Parse-8                          7.38µs ± 1%    6.07µs ± 2%  -17.70%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseReader-8                    10.1µs ± 1%     8.5µs ± 3%  -15.77%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Sign-8                           1.00ms ± 2%    0.96ms ± 1%   -4.02%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseString-8                    8.60µs ± 0%    7.53µs ± 1%  -12.54%  (p=0.016 n=4+5)
JWT/Serialization/JSON/jwt.Parse-8                          8.26µs ± 1%    7.51µs ± 2%   -9.14%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseReader-8                    11.1µs ± 1%    10.1µs ± 2%   -9.19%  (p=0.008 n=5+5)
JWT/Serialization/JSON/json.Unmarshal-8                      591ns ± 2%     599ns ± 3%     ~     (p=0.310 n=5+5)
JWT/Serialization/JSON/json.Marshal-8                       2.98µs ± 2%    2.96µs ± 3%     ~     (p=0.690 n=5+5)

name                                                      old alloc/op   new alloc/op   delta
JWE/Serialization/JSON/json.Marshal-8                       5.89kB ± 0%    5.84kB ± 0%   -0.83%  (p=0.000 n=4+5)
JWE/Serialization/JSON/json.Unmarshal-8                     1.58kB ± 0%    1.58kB ± 0%     ~     (all equal)
JWK/Serialization/RSA/PublicKey/jwk.Parse-8                 7.99kB ± 0%    7.16kB ± 0%  -10.41%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseString-8           8.41kB ± 0%    7.58kB ± 0%   -9.90%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseReader-8           7.95kB ± 0%    7.11kB ± 0%  -10.56%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/json.Marshal-8              2.51kB ± 0%    2.43kB ± 0%   -3.18%  (p=0.000 n=4+5)
JWK/Serialization/RSA/PrivateKey/jwk.Parse-8                48.0kB ± 0%    36.4kB ± 0%     ~     (p=0.079 n=4+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseString-8          49.7kB ± 0%    38.2kB ± 0%     ~     (p=0.079 n=4+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseReader-8          47.9kB ± 0%    36.4kB ± 0%  -24.04%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/json.Marshal-8             9.31kB ± 0%    9.03kB ± 0%   -2.95%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.Parse-8                  7.50kB ± 0%    6.97kB ± 0%   -7.04%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseString-8            7.74kB ± 0%    7.21kB ± 0%   -6.83%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseReader-8            7.45kB ± 0%    6.92kB ± 0%   -7.09%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/json.Marshal-8               1.99kB ± 0%    1.88kB ± 0%   -5.64%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.Parse-8                 8.06kB ± 0%    7.59kB ± 0%   -5.85%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseString-8           8.42kB ± 0%    7.94kB ± 0%   -5.61%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseReader-8           8.02kB ± 0%    7.54kB ± 0%   -5.89%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/json.Marshal-8              2.52kB ± 0%    2.40kB ± 0%   -5.11%  (p=0.000 n=5+4)
JWK/Serialization/Symmetric/PublicKey/jwk.Parse-8           6.91kB ± 0%    6.20kB ± 0%  -10.30%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseString-8     7.04kB ± 0%    6.33kB ± 0%  -10.11%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseReader-8     6.87kB ± 0%    6.15kB ± 0%  -10.48%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/json.Marshal-8        1.29kB ± 0%    1.23kB ± 0%   -4.97%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.Parse-8          6.91kB ± 0%    6.20kB ± 0%  -10.30%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseString-8    7.04kB ± 0%    6.33kB ± 0%  -10.11%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseReader-8    6.87kB ± 0%    6.15kB ± 0%  -10.48%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal-8       1.29kB ± 0%    1.23kB ± 0%   -4.97%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.Parse-8                       3.96kB ± 0%    2.81kB ± 0%  -29.09%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseString-8                 4.15kB ± 0%    3.00kB ± 0%  -27.75%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseReader-8                 4.48kB ± 0%    3.32kB ± 0%  -25.89%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.Parse-8                          21.5kB ± 0%    16.2kB ± 0%  -24.47%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseString-8                    22.4kB ± 0%    17.1kB ± 0%  -23.49%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseReader-8                    23.0kB ± 0%    17.8kB ± 0%  -22.84%  (p=0.008 n=5+5)
JWS/Serialization/JSON/json.Marshal-8                       6.10kB ± 0%    5.72kB ± 0%   -6.22%  (p=0.000 n=4+5)
JWT/Serialization/Sign/jwt.ParseString-8                    13.1kB ± 0%    10.9kB ± 0%  -16.76%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Parse-8                          12.7kB ± 0%    10.5kB ± 0%  -17.35%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseReader-8                    13.2kB ± 0%    11.0kB ± 0%  -16.68%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Sign-8                           36.4kB ± 0%    36.3kB ± 0%   -0.14%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseString-8                    11.2kB ± 0%    10.1kB ± 0%   -9.58%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.Parse-8                          11.1kB ± 0%    10.1kB ± 0%   -9.62%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseReader-8                    11.7kB ± 0%    10.6kB ± 0%   -9.06%  (p=0.008 n=5+5)
JWT/Serialization/JSON/json.Unmarshal-8                       576B ± 0%      576B ± 0%     ~     (all equal)
JWT/Serialization/JSON/json.Marshal-8                         817B ± 0%      800B ± 0%   -2.08%  (p=0.000 n=5+4)

name                                                      old allocs/op  new allocs/op  delta
JWE/Serialization/JSON/json.Marshal-8                         55.0 ± 0%      49.0 ± 0%  -10.91%  (p=0.008 n=5+5)
JWE/Serialization/JSON/json.Unmarshal-8                       26.0 ± 0%      26.0 ± 0%     ~     (all equal)
JWK/Serialization/RSA/PublicKey/jwk.Parse-8                   70.0 ± 0%      95.0 ± 0%  +35.71%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseString-8             71.0 ± 0%      96.0 ± 0%  +35.21%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseReader-8             69.0 ± 0%      94.0 ± 0%  +36.23%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/json.Marshal-8                33.0 ± 0%      24.0 ± 0%  -27.27%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.Parse-8                   135 ± 0%       204 ± 0%  +51.11%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseString-8             136 ± 0%       205 ± 0%  +50.74%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseReader-8             134 ± 0%       203 ± 0%  +51.49%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/json.Marshal-8               78.0 ± 0%      51.0 ± 0%  -34.62%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.Parse-8                    78.0 ± 0%     111.0 ± 0%  +42.31%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseString-8              79.0 ± 0%     112.0 ± 0%  +41.77%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseReader-8              77.0 ± 0%     110.0 ± 0%  +42.86%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/json.Marshal-8                 38.0 ± 0%      27.0 ± 0%  -28.95%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.Parse-8                   87.0 ± 0%     128.0 ± 0%  +47.13%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseString-8             88.0 ± 0%     129.0 ± 0%  +46.59%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseReader-8             86.0 ± 0%     127.0 ± 0%  +47.67%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/json.Marshal-8                45.0 ± 0%      31.0 ± 0%  -31.11%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.Parse-8             63.0 ± 0%      80.0 ± 0%  +26.98%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseString-8       64.0 ± 0%      81.0 ± 0%  +26.56%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseReader-8       62.0 ± 0%      79.0 ± 0%  +27.42%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/json.Marshal-8          26.0 ± 0%      20.0 ± 0%  -23.08%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.Parse-8            63.0 ± 0%      80.0 ± 0%  +26.98%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseString-8      64.0 ± 0%      81.0 ± 0%  +26.56%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseReader-8      62.0 ± 0%      79.0 ± 0%  +27.42%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal-8         26.0 ± 0%      20.0 ± 0%  -23.08%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.Parse-8                         43.0 ± 0%      48.0 ± 0%  +11.63%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseString-8                   44.0 ± 0%      49.0 ± 0%  +11.36%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseReader-8                   44.0 ± 0%      49.0 ± 0%  +11.36%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.Parse-8                             163 ± 0%       149 ± 0%   -8.59%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseString-8                       164 ± 0%       150 ± 0%   -8.54%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseReader-8                       165 ± 0%       151 ± 0%   -8.48%  (p=0.008 n=5+5)
JWS/Serialization/JSON/json.Marshal-8                         77.0 ± 0%      60.0 ± 0%  -22.08%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseString-8                      93.0 ± 0%     103.0 ± 0%  +10.75%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Parse-8                            92.0 ± 0%     102.0 ± 0%  +10.87%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseReader-8                      93.0 ± 0%     103.0 ± 0%  +10.75%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Sign-8                              195 ± 0%       187 ± 0%   -4.10%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseString-8                      69.0 ± 0%      74.0 ± 0%   +7.25%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.Parse-8                            68.0 ± 0%      73.0 ± 0%   +7.35%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseReader-8                      69.0 ± 0%      74.0 ± 0%   +7.25%  (p=0.008 n=5+5)
JWT/Serialization/JSON/json.Unmarshal-8                       10.0 ± 0%      10.0 ± 0%     ~     (all equal)
JWT/Serialization/JSON/json.Marshal-8                         22.0 ± 0%      18.0 ± 0%  -18.18%  (p=0.008 n=5+5)
```