@lestrrat 2021-01-24
For the record, if we use github.com/goccy/go-json, we currently have the following, which results in a whopping reduction of 70% in some cases

```
Benchmark comparison for:
  develop/v1.1.0 (curcommit)
  topic/v1.1.0-marshal-optimize-goccy (curcommit)

name                                                      old time/op    new time/op    delta
JWE/Serialization/JSON/json.Marshal-8                       13.1µs ± 1%    13.1µs ± 1%      ~     (p=0.421 n=5+5)
JWE/Serialization/JSON/json.Unmarshal-8                     3.38µs ± 1%    3.41µs ± 1%      ~     (p=0.095 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.Parse-8                 14.5µs ± 1%     4.7µs ± 3%   -67.81%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseString-8           14.5µs ± 1%     4.7µs ± 3%   -67.38%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseReader-8           16.9µs ± 1%     5.8µs ± 1%   -65.52%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/json.Marshal-8              6.11µs ± 4%    5.45µs ± 1%   -10.78%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.Parse-8                55.3µs ± 1%    13.9µs ± 8%   -74.84%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseString-8          55.3µs ± 2%    14.8µs ± 7%   -73.26%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseReader-8          58.7µs ± 0%    15.9µs ± 9%   -72.95%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/json.Marshal-8             19.1µs ± 2%    16.4µs ± 1%   -13.91%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.Parse-8                  11.5µs ± 1%     3.9µs ± 2%   -66.01%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseString-8            11.6µs ± 3%     4.0µs ± 2%   -65.51%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseReader-8            13.9µs ± 2%     5.5µs ± 4%   -60.72%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/json.Marshal-8               6.53µs ± 2%    5.93µs ± 2%    -9.07%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.Parse-8                 14.8µs ± 3%     4.8µs ± 2%   -67.67%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseString-8           14.7µs ± 1%     4.8µs ± 2%   -67.42%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseReader-8           17.4µs ± 2%     6.1µs ± 2%   -64.89%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/json.Marshal-8              7.93µs ± 1%    7.25µs ± 3%    -8.56%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.Parse-8           7.64µs ± 3%    2.91µs ± 6%   -61.97%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseString-8     7.62µs ± 1%    2.91µs ± 2%   -61.78%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseReader-8     9.41µs ± 1%    4.37µs ± 2%   -53.53%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/json.Marshal-8        3.59µs ± 1%    3.55µs ± 4%      ~     (p=0.222 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.Parse-8          7.39µs ± 1%    2.86µs ± 3%   -61.33%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseString-8    7.43µs ± 1%    2.90µs ± 1%   -61.00%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseReader-8    9.45µs ± 1%    4.42µs ± 3%   -53.18%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal-8       3.61µs ± 3%    3.92µs ±16%      ~     (p=0.310 n=5+5)
JWS/Serialization/Compact/jws.Parse-8                       3.15µs ± 1%    1.40µs ± 6%   -55.63%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseString-8                 3.20µs ± 1%    1.42µs ± 1%   -55.76%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseReader-8                 5.02µs ± 1%    3.10µs ± 3%   -38.30%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.Parse-8                          21.0µs ± 1%     7.3µs ± 4%   -65.27%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseString-8                    21.3µs ± 1%     7.5µs ± 2%   -64.90%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseReader-8                    24.7µs ± 1%     9.5µs ± 2%   -61.65%  (p=0.008 n=5+5)
JWS/Serialization/JSON/json.Marshal-8                       13.0µs ± 2%    11.2µs ± 1%   -13.57%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseString-8                    7.45µs ± 1%    3.56µs ± 2%   -52.17%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Parse-8                          7.37µs ± 1%    3.45µs ± 1%   -53.22%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseReader-8                    9.86µs ± 1%    5.53µs ± 1%   -43.97%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Sign-8                            990µs ± 2%     982µs ± 2%      ~     (p=0.310 n=5+5)
JWT/Serialization/JSON/jwt.ParseString-8                    8.18µs ± 1%    4.48µs ± 1%   -45.25%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.Parse-8                          8.16µs ± 1%    4.44µs ± 1%   -45.55%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseReader-8                    10.8µs ± 1%     6.5µs ± 1%      ~     (p=0.095 n=5+2)

name                                                      old alloc/op   new alloc/op   delta
JWE/Serialization/JSON/json.Marshal-8                       5.89kB ± 0%   12.83kB ± 0%  +117.84%  (p=0.016 n=4+5)
JWE/Serialization/JSON/json.Unmarshal-8                     1.58kB ± 0%    1.58kB ± 0%      ~     (all equal)
JWK/Serialization/RSA/PublicKey/jwk.Parse-8                 7.99kB ± 0%    8.54kB ± 0%    +6.83%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseString-8           8.41kB ± 0%    8.95kB ± 0%    +6.49%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseReader-8           7.95kB ± 0%    8.50kB ± 0%    +6.84%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/json.Marshal-8              2.51kB ± 0%    4.73kB ± 0%   +88.81%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.Parse-8                48.0kB ± 0%    26.0kB ± 0%      ~     (p=0.079 n=4+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseString-8          49.8kB ± 0%    27.7kB ± 0%   -44.23%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseReader-8          47.9kB ± 0%    25.9kB ± 0%   -45.93%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/json.Marshal-8             9.31kB ± 0%   11.34kB ± 0%   +21.82%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.Parse-8                  7.50kB ± 0%    7.80kB ± 0%    +4.06%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseString-8            7.74kB ± 0%    8.04kB ± 0%    +3.93%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseReader-8            7.45kB ± 0%    7.76kB ± 0%    +4.19%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/json.Marshal-8               1.99kB ± 0%    4.18kB ± 0%  +110.52%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.Parse-8                 8.06kB ± 0%    8.55kB ± 0%    +6.05%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseString-8           8.42kB ± 0%    8.90kB ± 0%    +5.80%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseReader-8           8.02kB ± 0%    8.51kB ± 0%    +6.19%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/json.Marshal-8              2.52kB ± 0%    4.70kB ± 0%   +86.37%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.Parse-8           6.91kB ± 0%    7.18kB ± 0%    +3.94%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseString-8     7.04kB ± 0%    7.31kB ± 0%    +3.86%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseReader-8     6.87kB ± 0%    7.14kB ± 0%    +3.84%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/json.Marshal-8        1.29kB ± 0%    3.53kB ± 0%  +174.01%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.Parse-8          6.91kB ± 0%    7.18kB ± 0%    +3.94%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseString-8    7.04kB ± 0%    7.31kB ± 0%    +3.86%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseReader-8    6.87kB ± 0%    7.14kB ± 0%    +3.84%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal-8       1.29kB ± 0%    3.53kB ± 0%  +174.04%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.Parse-8                       3.96kB ± 0%    3.00kB ± 0%   -24.24%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseString-8                 4.15kB ± 0%    3.19kB ± 0%   -23.12%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseReader-8                 4.48kB ± 0%    3.51kB ± 0%   -21.61%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.Parse-8                          21.5kB ± 0%    15.9kB ± 0%   -26.20%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseString-8                    22.4kB ± 0%    16.8kB ± 0%   -25.15%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseReader-8                    23.0kB ± 0%    17.4kB ± 0%   -24.38%  (p=0.008 n=5+5)
JWS/Serialization/JSON/json.Marshal-8                       6.10kB ± 0%   14.95kB ± 0%  +145.18%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseString-8                    13.1kB ± 0%     8.9kB ± 0%   -32.24%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Parse-8                          12.7kB ± 0%     8.4kB ± 0%   -33.38%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseReader-8                    13.2kB ± 0%     9.0kB ± 0%   -32.08%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Sign-8                           36.4kB ± 0%    41.0kB ± 0%   +12.87%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseString-8                    11.2kB ± 0%     8.5kB ± 0%   -24.16%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.Parse-8                          11.1kB ± 0%     8.4kB ± 0%   -24.26%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseReader-8                    11.7kB ± 0%     9.0kB ± 0%   -23.20%  (p=0.000 n=5+2)

name                                                      old allocs/op  new allocs/op  delta
JWE/Serialization/JSON/json.Marshal-8                         55.0 ± 0%      62.0 ± 0%   +12.73%  (p=0.008 n=5+5)
JWE/Serialization/JSON/json.Unmarshal-8                       26.0 ± 0%      26.0 ± 0%      ~     (all equal)
JWK/Serialization/RSA/PublicKey/jwk.Parse-8                   70.0 ± 0%      53.0 ± 0%   -24.29%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseString-8             71.0 ± 0%      54.0 ± 0%   -23.94%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/jwk.ParseReader-8             69.0 ± 0%      52.0 ± 0%   -24.64%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PublicKey/json.Marshal-8                33.0 ± 0%      29.0 ± 0%   -12.12%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.Parse-8                   135 ± 0%        93 ± 0%   -31.11%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseString-8             136 ± 0%        94 ± 0%   -30.88%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/jwk.ParseReader-8             134 ± 0%        92 ± 0%   -31.34%  (p=0.008 n=5+5)
JWK/Serialization/RSA/PrivateKey/json.Marshal-8               78.0 ± 0%      56.0 ± 0%   -28.21%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.Parse-8                    78.0 ± 0%      58.0 ± 0%   -25.64%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseString-8              79.0 ± 0%      59.0 ± 0%   -25.32%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/jwk.ParseReader-8              77.0 ± 0%      57.0 ± 0%   -25.97%  (p=0.008 n=5+5)
JWK/Serialization/EC/PublicKey/json.Marshal-8                 38.0 ± 0%      32.0 ± 0%   -15.79%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.Parse-8                   87.0 ± 0%      65.0 ± 0%   -25.29%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseString-8             88.0 ± 0%      66.0 ± 0%   -25.00%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/jwk.ParseReader-8             86.0 ± 0%      64.0 ± 0%   -25.58%  (p=0.008 n=5+5)
JWK/Serialization/EC/PrivateKey/json.Marshal-8                45.0 ± 0%      36.0 ± 0%   -20.00%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.Parse-8             63.0 ± 0%      49.0 ± 0%   -22.22%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseString-8       64.0 ± 0%      50.0 ± 0%   -21.88%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/jwk.ParseReader-8       62.0 ± 0%      48.0 ± 0%   -22.58%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PublicKey/json.Marshal-8          26.0 ± 0%      25.0 ± 0%    -3.85%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.Parse-8            63.0 ± 0%      49.0 ± 0%   -22.22%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseString-8      64.0 ± 0%      50.0 ± 0%   -21.88%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/jwk.ParseReader-8      62.0 ± 0%      48.0 ± 0%   -22.58%  (p=0.008 n=5+5)
JWK/Serialization/Symmetric/PrivateKey/json.Marshal-8         26.0 ± 0%      25.0 ± 0%    -3.85%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.Parse-8                         43.0 ± 0%      28.0 ± 0%   -34.88%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseString-8                   44.0 ± 0%      29.0 ± 0%   -34.09%  (p=0.008 n=5+5)
JWS/Serialization/Compact/jws.ParseReader-8                   44.0 ± 0%      29.0 ± 0%   -34.09%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.Parse-8                             163 ± 0%       103 ± 0%   -36.81%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseString-8                       164 ± 0%       104 ± 0%   -36.59%  (p=0.008 n=5+5)
JWS/Serialization/JSON/jws.ParseReader-8                       165 ± 0%       105 ± 0%   -36.36%  (p=0.008 n=5+5)
JWS/Serialization/JSON/json.Marshal-8                         77.0 ± 0%      80.0 ± 0%    +3.90%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseString-8                      93.0 ± 0%      67.0 ± 0%   -27.96%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Parse-8                            92.0 ± 0%      66.0 ± 0%   -28.26%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.ParseReader-8                      93.0 ± 0%      67.0 ± 0%   -27.96%  (p=0.008 n=5+5)
JWT/Serialization/Sign/jwt.Sign-8                              195 ± 0%       197 ± 0%    +1.03%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseString-8                      69.0 ± 0%      59.0 ± 0%   -14.49%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.Parse-8                            68.0 ± 0%      58.0 ± 0%   -14.71%  (p=0.008 n=5+5)
JWT/Serialization/JSON/jwt.ParseReader-8                      69.0 ± 0%      59.0 ± 0%   -14.49%  (p=0.000 n=5+2)
```